# backend/src/swagger.json
{
	"components": {
...
		"schemas": {
			"ErrorResponse": {
				"description": "Error response.",
				"properties": {
					"responseCode": {
						"type": "number",
						"format": "double",
						"description": "error code in system."
					},
					"message": {
						"type": "string",
						"description": "Readable message."
					}
				},
				"required": [
					"responseCode"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SwitchOptions": {
				"type": "string",
				"enum": [
					"on",
					"off"
				],
				"description": "Switches option"
			},
...
			"MinionStatus": {
				"description": "Minion status, the available values depend on the minion type.",
				"properties": {
					"toggle": {
						"$ref": "#/components/schemas/Toggle"
					},
					"switch": {
						"$ref": "#/components/schemas/Switch"
					},
					"roller": {
						"$ref": "#/components/schemas/Roller"
					},
					"cleaner": {
						"$ref": "#/components/schemas/Cleaner"
					},
					"airConditioning": {
						"$ref": "#/components/schemas/AirConditioning"
					},
					"light": {
						"$ref": "#/components/schemas/Light"
					},
					"temperatureLight": {
						"$ref": "#/components/schemas/TemperatureLight"
					},
					"colorLight": {
						"$ref": "#/components/schemas/ColorLight"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"CalibrationMode": {
				"type": "string",
				"enum": [
					"LOCK_ON",
					"LOCK_OFF",
					"SHABBAT",
					"AUTO"
				]
			},
			"MinionCalibrate": {
				"description": "Used to change the minion calibration property value.",
				"properties": {
					"calibrationCycleMinutes": {
						"type": "integer",
						"format": "int32",
						"description": "Minutes to calibrate status, set 0 to turn off calibration",
						"minimum": 0
					},
					"calibrationMode": {
						"$ref": "#/components/schemas/CalibrationMode",
						"description": "The calibration mode to calibrate"
					}
				},
				"required": [
					"calibrationCycleMinutes",
					"calibrationMode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Minion": {
				"description": "Represents a minion in system.\r\nMinion is a logic device in the system, meaning that a device is a physical device and minion is a logic device\r\nthat uses a physical device to switch home kit status. For example, an IR transmitter can be one physical device\r\nfor a few minions, one to central AC control and second for secondary AC control\r\nso in it will be two totally different minions that use one physical device.",
				"properties": {
					"minionId": {
						"type": "string",
						"description": "Minion unique id."
					},
					"name": {
						"type": "string",
						"description": "The display name for a minion."
					},
					"device": {
						"$ref": "#/components/schemas/MinionDevice",
						"description": "Physical device of minion."
					},
					"isProperlyCommunicated": {
						"type": "boolean",
						"description": "Is communication with device status ok."
					},
					"minionStatus": {
						"$ref": "#/components/schemas/MinionStatus",
						"description": "Status of minion (based on minion type)."
					},
					"minionType": {
						"$ref": "#/components/schemas/MinionTypes",
						"description": "Minion type."
					},
					"minionAutoTurnOffMS": {
						"type": "number",
						"format": "double",
						"description": "Auto turns  off duration, *if* set member value then the minion will turn off in X ms after turning it on,\r\nUsed for example in boiler minion etc."
					},
					"calibration": {
						"$ref": "#/components/schemas/MinionCalibrate",
						"description": "Calibrate the physical device with the server known status, in a periodic cycle,\r\nand allow locking the status."
					},
					"room": {
						"type": "string",
						"description": "Represents the room where the minion is located at."
					}
				},
				"required": [
					"name",
					"device",
					"minionStatus",
					"minionType"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IftttActionTriggered": {
				"description": "Ifttt trigger set status action",
				"properties": {
					"apiKey": {
						"type": "string",
						"description": "WebHooks API key"
					},
					"localMac": {
						"type": "string",
						"description": "Allow remote-server to forward request to local server"
					},
					"setStatus": {
						"$ref": "#/components/schemas/SwitchOptions"
					}
				},
				"required": [
					"apiKey",
					"setStatus"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"MinionTimeline": {
				"description": "Minion timeline node",
				"properties": {
					"minionId": {
						"type": "string"
					},
					"timestamp": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/MinionStatus"
					},
					"trigger": {
						"$ref": "#/components/schemas/MinionChangeTrigger"
					},
					"user": {
						"properties": {
							"email": {
								"type": "string"
							},
							"name": {
								"type": "string"
							}
						},
						"required": [
							"email",
							"name"
						],
						"type": "object"
					}
				},
				"required": [
					"minionId",
					"timestamp",
					"status",
					"trigger"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MinionRename": {
				"description": "Used to rename minion.",
				"properties": {
					"name": {
						"type": "string",
						"description": "The new name to set."
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MinionSetRoomName": {
				"description": "Used to set minion room name.",
				"properties": {
					"room": {
						"type": "string",
						"description": "The new room name to set."
					}
				},
				"required": [
					"room"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SetMinionAutoTurnOff": {
				"description": "Used to change the minion auto turn off the timeout value.",
				"properties": {
					"setAutoTurnOffMS": {
						"type": "number",
						"format": "double",
						"description": "The timeout duration in ms, to turn off set 0 as valse."
					}
				},
				"required": [
					"setAutoTurnOffMS"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ScanningStatus": {
				"description": "Scanning progress status",
				"properties": {
					"scanningStatus": {
						"$ref": "#/components/schemas/ProgressStatus"
					}
				},
				"required": [
					"scanningStatus"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"userAuth": {
				"type": "apiKey",
				"name": "authentication",
				"in": "header"
			},
			"adminAuth": {
				"type": "apiKey",
				"name": "authentication",
				"in": "header"
			},
			"iftttAuth": {
				"type": "apiKey",
				"name": "apiKey",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "casanet-local-server",
		"version": "2.0.0",
		"description": "IoT server for local home network - Server side",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Haim Kastner",
			"email": "haim.kastner@gmail.com",
			"url": "https://github.com/casanet/casanet-server"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/ifttt/trigger/minions/{minionId}": {
			"post": {
				"operationId": "TriggeredMinionAction",
				"responses": {
					"204": {
						"description": "No content"
					},
					"501": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Ifttt webhooks triggering casa-net *minion* action API.",
				"tags": [
					"Ifttt"
				],
				"deprecated": true,
				"security": [
					{
						"iftttAuth": []
					}
				],
				"parameters": [
					{
						"description": "minion to set status.",
						"in": "path",
						"name": "minionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "status to set.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IftttActionTriggered"
							}
						}
					}
				}
			}
		},
...
		"/minions/timeline": {
			"get": {
				"operationId": "GetMinionsTimeline",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MinionTimeline"
									},
									"type": "array"
								}
							}
						}
					},
					"501": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get the timeline of minions status.",
				"tags": [
					"Minions"
				],
				"security": [
					{
						"userAuth": []
					},
					{
						"adminAuth": []
					}
				],
				"parameters": []
			}
		},
		"/minions/timeline/{minionId}": {
			"get": {
				"operationId": "GetMinionTimeline",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MinionTimeline"
									},
									"type": "array"
								}
							}
						}
					},
					"501": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get the timeline of minions status.",
				"tags": [
					"Minions"
				],
				"security": [
					{
						"userAuth": []
					},
					{
						"adminAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "minionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/minions/power-off": {
			"put": {
				"operationId": "PowerAllOff",
				"responses": {
					"204": {
						"description": "No content"
					},
					"501": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Power off all minions",
				"tags": [
					"Minions"
				],
				"security": [
					{
						"userAuth": []
					},
					{
						"adminAuth": []
					}
				],
				"parameters": []
			}
		},
		"/minions/rename/{minionId}": {
			"put": {
				"operationId": "RenameMinion",
				"responses": {
					"204": {
						"description": "No content"
					},
					"501": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Update minion name.",
				"tags": [
					"Minions"
				],
				"security": [
					{
						"userAuth": []
					},
					{
						"adminAuth": []
					}
				],
				"parameters": [
					{
						"description": "Minion id.",
						"in": "path",
						"name": "minionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MinionRename"
							}
						}
					}
				}
			}
		},
		"/minions/bluetooth/rename/{minionId}": {
			"put": {
				"operationId": "RenameBluetoothMinion",
				"responses": {
					"204": {
						"description": "No content"
					},
					"501": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Update minion name.",
				"tags": [
					"Minions"
				],
				"security": [
					{
						"userAuth": []
					},
					{
						"adminAuth": []
					}
				],
				"parameters": [
					{
						"description": "Minion id.",
						"in": "path",
						"name": "minionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MinionRename"
							}
						}
					}
				}
			}
		},
		"/minions/room/{minionId}": {
			"put": {
				"operationId": "RenameRoom",
				"responses": {
					"204": {
						"description": "No content"
					},
					"501": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Update minion room name.",
				"tags": [
					"Minions"
				],
				"security": [
					{
						"userAuth": []
					},
					{
						"adminAuth": []
					}
				],
				"parameters": [
					{
						"description": "Minion id.",
						"in": "path",
						"name": "minionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Minion room name to set.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MinionSetRoomName"
							}
						}
					}
				}
			}
		},
		"/minions/timeout/{minionId}": {
			"put": {
				"operationId": "SetMinionTimeout",
				"responses": {
					"204": {
						"description": "No content"
					},
					"501": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Update minion auto turns off timeout.",
				"tags": [
					"Minions"
				],
				"security": [
					{
						"userAuth": []
					},
					{
						"adminAuth": []
					}
				],
				"parameters": [
					{
						"description": "Minion id.",
						"in": "path",
						"name": "minionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Timeout property.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SetMinionAutoTurnOff"
							}
						}
					}
				}
			}
		},
		"/minions/calibrate/{minionId}": {
			"put": {
				"operationId": "SetMinionCalibrate",
				"responses": {
					"204": {
						"description": "No content"
					},
					"501": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Update minion auto turns off timeout.",
				"tags": [
					"Minions"
				],
				"security": [
					{
						"userAuth": []
					},
					{
						"adminAuth": []
					}
				],
				"parameters": [
					{
						"description": "Minion id.",
						"in": "path",
						"name": "minionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MinionCalibrate"
							}
						}
					}
				}
			}
		},
		"/minions/rescan/{minionId}": {
			"post": {
				"operationId": "RescanMinionStatus",
				"responses": {
					"204": {
						"description": "No content"
					},
					"501": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Recheck minion device status (update server status cache).",
				"tags": [
					"Minions"
				],
				"security": [
					{
						"userAuth": []
					},
					{
						"adminAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "minionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/minions/rescan": {
			"post": {
				"operationId": "RescanMinionsStatus",
				"responses": {
					"204": {
						"description": "No content"
					},
					"501": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Recheck every minion device status (update server status cache).\r\nNote that this is not the devices scan!\r\nThis scanning only checks every minion API to know the current status.",
				"tags": [
					"Minions"
				],
				"security": [
					{
						"userAuth": []
					},
					{
						"adminAuth": []
					}
				],
				"parameters": [
					{
						"description": "Whenever scan also the local networks IP's map or not",
						"in": "query",
						"name": "scanNetwork",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					}
				]
			},
			"get": {
				"operationId": "GetSescaningMinionsStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ScanningStatus"
								}
							}
						}
					},
					"501": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get the current scanning status",
				"tags": [
					"Minions"
				],
				"security": [
					{
						"userAuth": []
					},
					{
						"adminAuth": []
					}
				],
				"parameters": []
			}
		},
		"/minions/{minionId}": {
			"delete": {
				"operationId": "DeleteMinion",
				"responses": {
					"204": {
						"description": "No content"
					},
					"501": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Delete minion from the system.",
				"tags": [
					"Minions"
				],
				"security": [
					{
						"userAuth": []
					},
					{
						"adminAuth": []
					}
				],
				"parameters": [
					{
						"description": "Minion id.",
						"in": "path",
						"name": "minionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"get": {
				"operationId": "GetMinion",
				"responses": {
					"200": {
						"description": "Minion.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Minion"
								}
							}
						}
					}
				},
				"description": "Get minion by id.",
				"tags": [
					"Minions"
				],
				"security": [
					{
						"userAuth": []
					},
					{
						"adminAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "minionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "SetMinion",
				"responses": {
					"204": {
						"description": "No content"
					},
					"501": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Update minion status.",
				"tags": [
					"Minions"
				],
				"security": [
					{
						"userAuth": []
					},
					{
						"adminAuth": []
					}
				],
				"parameters": [
					{
						"description": "Minion id.",
						"in": "path",
						"name": "minionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Minion status to set.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MinionStatus"
							}
						}
					}
				}
			}
		},
		"/minions/bluetooth/{minionId}": {
			"delete": {
				"operationId": "DeleteBluetoothMinion",
				"responses": {
					"204": {
						"description": "No content"
					},
					"501": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Delete minion from the system.",
				"tags": [
					"Minions"
				],
				"security": [
					{
						"userAuth": []
					},
					{
						"adminAuth": []
					}
				],
				"parameters": [
					{
						"description": "Minion id.",
						"in": "path",
						"name": "minionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"get": {
				"operationId": "GetBluetoothMinion",
				"responses": {
					"200": {
						"description": "Minion.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BluetoothMinion"
								}
							}
						}
					}
				},
				"description": "Get minion by id.",
				"tags": [
					"Minions"
				],
				"security": [
					{
						"userAuth": []
					},
					{
						"adminAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "minionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/minions": {
			"post": {
				"operationId": "CreateMinion",
				"responses": {
					"204": {
						"description": "No content"
					},
					"501": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Creates a new minion.",
				"tags": [
					"Minions"
				],
				"security": [
					{
						"userAuth": []
					},
					{
						"adminAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "The new minion to create.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Minion"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetMinions",
				"responses": {
					"200": {
						"description": "Minions array.",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Minion"
									},
									"type": "array"
								}
							}
						}
					},
					"501": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get all minions in the system.",
				"tags": [
					"Minions"
				],
				"security": [
					{
						"userAuth": []
					},
					{
						"adminAuth": []
					}
				],
				"parameters": []
			}
		},
		"/minions/bluetooth": {
			"post": {
				"operationId": "CreateBluetoothMinion",
				"responses": {
					"204": {
						"description": "No content"
					},
					"501": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Creates a new minion.",
				"tags": [
					"Minions"
				],
				"security": [
					{
						"userAuth": []
					},
					{
						"adminAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "The new minion to create.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BluetoothMinion"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetBluetoothMinions",
				"responses": {
					"200": {
						"description": "Minions array.",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/BluetoothMinion"
									},
									"type": "array"
								}
							}
						}
					},
					"501": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get all minions in the system.",
				"tags": [
					"Minions"
				],
				"security": [
					{
						"userAuth": []
					},
					{
						"adminAuth": []
					}
				],
				"parameters": []
			}
		},
		"/minions/{minionId}/ifttt": {
			"put": {
				"operationId": "NotifyMinionStatusChanged",
				"responses": {
					"204": {
						"description": "No content"
					},
					"501": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Notify minion status changed by ifttt webhook (https://ifttt.com/maker_webhooks).",
				"tags": [
					"Minions"
				],
				"security": [
					{
						"iftttAuth": []
					}
				],
				"parameters": [
					{
						"description": "Minion id.",
						"in": "path",
						"name": "minionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Minion key amd status to set.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IftttOnChanged"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/API"
		}
	]
}